requirements.txt
langchain-google-genai>=0.0.8
pydantic>=2.0.0
python-dotenv>=1.0.0